requirements.txt
librosa
matplotlib == 3.9.0
numba == 0.59.1
numpy == 1.26.4
PyYAML
scienceplots
scipy
soundfile
